test(notes): add integration tests for loadConfig and pipeline behavior
- Introduced integration tests for loadConfig to verify behavior with various config scenarios, including handling of environment variables and JSONC parsing.
- Added tests for pipeline configuration to ensure correct passing of LLM options and handling of changelog modes.
- Removed outdated unit tests from core.spec.ts to streamline test coverage and focus on integration testing.
